Convert to float16 in arrow conversion
llama2 weights are bfloat16 and currently these are not supported by numpy

+++ llama/generation.py
@@ -102,7 +102,7 @@ class Llama:
                 ckpt = torch.load(ckpt_file, map_location="cpu")
                 (arrow_dir / index).mkdir(parents=True, exist_ok=True)
                 for k, v in ckpt.items():
-                    tens = pa.Tensor.from_numpy(v.numpy())
+                    tens = pa.Tensor.from_numpy(v.to(torch.float16).numpy())
                     with pa.output_stream(arrow_dir / index / k) as f:
                         pa.ipc.write_tensor(tens, f)
                 ckpt = None
